+++ packages/nimbus/configs/babel.js
@@ -4,9 +4,7 @@ const { getConfig } = require('@rajzik/config-babel');
 const { getSettings } = require('@rajzik/nimbus-common');
 
 const { context, tool } = process.beemo;
-const {
-  graphql, library, next, node, react, env,
-} = getSettings();
+const { graphql, library, next, node, react, env } = getSettings();
 
 module.exports = getConfig({
   env,