+++ packages/nimbus/configs/jest.js
@@ -3,9 +3,7 @@
 const { getConfig } = require('@rajzik/config-jest');
 const { getSettings } = require('@rajzik/nimbus-common');
 
-const {
-  coverage, graphql, react, srcFolder, testFolder,
-} = getSettings();
+const { coverage, graphql, react, srcFolder, testFolder } = getSettings();
 
 module.exports = getConfig({
   srcFolder,