+++ packages/config-babel/CHANGELOG.md
@@ -7,8 +7,8 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 #### 📦 Dependencies
 
-- **[Babel]** Update to latest. ([755cf98](https://github.com/airbnb/nimbus/commit/755cf98))
-- **[Beemo]** Update to latest. ([bf6798c](https://github.com/airbnb/nimbus/commit/bf6798c))
+- **[Babel]** Update to latest. ([755cf98](https://github.com/rajzik/nimbus/commit/755cf98))
+- **[Beemo]** Update to latest. ([bf6798c](https://github.com/rajzik/nimbus/commit/bf6798c))
 
 **Note:** Version bump only for package @rajzik/config-babel
 
@@ -20,7 +20,7 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 #### 📦 Dependencies
 
-- **[Babel]** Update to v7.6.2. ([b657a49](https://github.com/airbnb/nimbus/commit/b657a49))
+- **[Babel]** Update to v7.6.2. ([b657a49](https://github.com/rajzik/nimbus/commit/b657a49))
 
 **Note:** Version bump only for package @rajzik/config-babel